# application/services/custom_services/testapp/src/models/entitymodels/testapp.ts
export interface testapp 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   name: String,
   test: String
}